renamed jenkins.py and its service to ci
